Adiciona comentário nos caixas diários

+++ database/migrations/2026_01_06_001538_add_preferences_to_user_table.php
@@ -4,25 +4,23 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
+return new class extends Migration {
+    public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->json('preferences')->nullable()->after('password');
+            if (!Schema::hasColumn('users', 'preferences')) {
+                $table->json('preferences')->nullable()->after('password');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
-    public function down(): void
+    public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('preferences');
+            if (Schema::hasColumn('users', 'preferences')) {
+                $table->dropColumn('preferences');
+            }
         });
     }
 };
+
